Security enhanced Diveclub/user relation modified Changed multiple divesite characteristics User can't modify certificate anymore Restricted access to pages for non admins Changed delete to only a flag in the db

+++ user/user_diveclub_delete.php
@@ -1,6 +1,16 @@
 <?php
 include "../function.php";
-$stmt = "DELETE FROM diveclub_user WHERE id_user = ? AND id_diveclub = ?";
-$pdo->prepare($stmt)->execute([$_GET['id'], $_GET['diveclub']]);
-$_SESSION['success'] = "Duikclub " . $_GET['diveclub'] . " werd succesvol verwijderd.";
+isAdminOrInstructorRedirect();
+$stmt = $pdo->prepare("SELECT count(*) FROM diveclub_user WHERE id_user = ?");
+$stmt->execute(array($_GET['id']));
+$currentClubs = $stmt->fetchColumn();
+
+if ($currentClubs == 1) {
+    $_SESSION['error'] = "Duikclub verwijderen is mislukt, je moet minimaal 1 duikclub overhouden.";
+} else {
+    $stmt = "DELETE FROM diveclub_user WHERE id_user = ? AND id_diveclub = ?";
+    $pdo->prepare($stmt)->execute([$_GET['id'], $_GET['diveclub']]);
+    $_SESSION['success'] = "De duikclub werd succesvol verwijderd.";
+}
+
 header("location: user_diveclub.php?id=" . $_GET['id']);
